requirements.txt
autopep8==2.1.0
blinker==1.7.0
click==8.1.7
Flask==3.0.2
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.1
pep8==1.7.1
pycodestyle==2.11.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyyml==0.0.2
scikit-learn==1.3.0
scipy==1.12.0
six==1.16.0
threadpoolctl==3.4.0
tzdata==2024.1
Werkzeug==3.0.1
